requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
dnspython==2.4.2
validate-email-address==1.0
gunicorn==21.2.0